backend/pyproject.toml
[project]
name = "extrashifty"
version = "0.1.0"
description = "ExtraShifty - Shift Management Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlmodel>=0.0.22",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "pyjwt>=2.9.0",
    "pwdlib[argon2,bcrypt]>=0.2.0",
    "python-multipart>=0.0.18",  # SECURITY: CVE-2024-24762, CVE-2024-53981
    "email-validator>=2.2.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "emails>=0.6",
    "jinja2>=3.1.4",
    "slowapi>=0.1.9",
    "stripe>=10.0.0",
    "apscheduler>=3.10.0",  # For scheduled jobs
    "weasyprint>=62.0",  # For PDF invoice generation
    "sentry-sdk>=2.0",  # For error tracking and monitoring
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "coverage>=7.6.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
